added hook useApplication
