+++ src/components/shared/Application/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useApplication } from "./useApplication";
